Fixed wrong serial size in stlink_probe_usb_devs()
In stlink_probe_usb_devs(), we detect available st-link devices, reading their device serials.
We then call stlink_open_usb() with the serial as an argument to uniquely identify the device to
query its details (like st-link version). As stlink_probe_usb_devs() only reads 13 bytes of the
serial, stlink_open_usb() reads 16 bytes, the serial comparison fails and the device is not
detected correctly.

We workaround that problem by making both of the functions read the first 16 bytes of the serial.