Added support for STM32H743ZI-NUCLEO
As stm32h7 and f7 are pretty similar, the flashloaders are copied; might not
be necessary, but I didn't know if I had to adopt them in the (near) future.
This changed is staged for revert if there is no necessity to keep them.

Added new Chip-IDs for STM32H743ZI-NUCLEO. As a workarund, the DBGMCU->IDC
address was hardcoded besides the previous addresses in stlink_chip_id().